requirements.txt
fastapi
uvicorn
transformers
langchain
langchain_huggingface
huggingface_hub
python-dotenv
pydantic
redis
firebase-admin
requests
psycopg2